pyproject.toml
[project]
name = "civil-law-agent"
version = "0.1.0"
description = "民商事司法智能体 - 外商投资纠纷 MVP"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pywin32>=306",
    "whoosh>=2.7.4",
    "jieba>=0.42.1",
    "python-docx>=1.1.0",
    "chromadb>=0.5.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "gradio>=4.0.0",
]
